Changed count to uint8

+++ VendorPt3.sol
@@ -3,7 +3,7 @@ pragma solidity 0.4.25;
 contract VendorPt3 {
   struct Item {
     string name;
-    uint count;
+    uint8 count;
   }
 
   address storeOwner;
